pc/by - refactor to add an interface
